Isolated rest error handling from the page router into separate utility to be used on other routers
