+++ services/app-management-service/src/middleware/validator.ts
@@ -1,7 +1,7 @@
 import { plainToInstance } from 'class-transformer';
 import { validate as classValidator } from 'class-validator';
 import { RequestHandler } from 'express';
-import { BadRequestError } from './error';
+import { BadRequestError } from '../error/errors';
  
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const validate = (type: any): RequestHandler => {
